Provide AlertManager-specific configuration base64 string
This change allows the charm user to provide AlertManager configuration
as base64-encoded strings. The rationale for this is so that systems
administrators who want to re-use an existing alertmanager.yml file can do so.
This also avoids the "abstraction over an abstraction" problem that would
be present if AlertManager-specific configuration was defined via the
charm's config.yml file. For example, AlertManager's inhibit_rules config
would have to be hand-massaged by the charm user to fit whatever is defined in
the charm's config.yaml file.

Additionl conversation is archived at: https://bit.ly/3fh0qNJ