Initial work on dependency management
Note that this change is incomplete and will require changes to the code
such that it doesn't explicitly add the path to the ops lib but instead
just imports it under the assumption that the dependency has already
been resolved for it. How `poetry install` will be called at runtime is
outside the scope of this charm.